requirements.txt
flask==3.0.0
pymupdf==1.27.2.2
pillow==12.1.1
gunicorn==21.2.0
gradio==4.20.0